+++ projectroles/templatetags/projectroles_common_tags.py
@@ -7,7 +7,6 @@ from django import template
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.staticfiles import finders
-from django.core.exceptions import ObjectDoesNotExist
 from django.template.loader import get_template
 from django.templatetags.static import static
 from django.urls import reverse
@@ -16,7 +15,7 @@ from djangoplugins.models import Plugin
 import projectroles
 from projectroles.app_settings import AppSettingAPI
 from projectroles.models import Project, RemoteProject, SODAR_CONSTANTS
-from projectroles.plugins import get_backend_api, BackendPluginPoint
+from projectroles.plugins import get_backend_api
 from projectroles.utils import get_display_name as _get_display_name
 
 
@@ -203,15 +202,14 @@ def get_user_html(user):
 
 @register.simple_tag
 def get_backend_include(backend_name, include_type='js'):
-    """Returns import string for backend app Javascript or CSS.
-    Returns empty string if not found."""
+    """
+    Return import string for backend app Javascript or CSS.
+    Returns empty string if not found.
+    """
+    from projectroles.plugins import get_app_plugin
 
-    # TODO: Replace with get_app_plugin() and if None check
-    # TODO: once get_app_plugin() can be used for backend plugins
-    # TODO: Don't forget to remove ObjectDoesNotExist import
-    try:
-        plugin = BackendPluginPoint.get_plugin(backend_name)
-    except ObjectDoesNotExist:
+    plugin = get_app_plugin(backend_name, plugin_type='backend')
+    if not plugin:
         return ''
 
     include = ''
@@ -230,7 +228,6 @@ def get_backend_include(backend_name, include_type='js'):
 
     if include and finders.find(include):
         return include_string.format(static(include))
-
     return ''