+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 
 - **Appalerts**
     - Update alerts in JQuery without page reloading (#701)
+    - ``APPALERTS_STATUS_INTERVAL`` setting (#701)
 
 Changed
 -------
@@ -24,6 +25,7 @@ Changed
     - Upgrade django-debug-toolbar to v3.2.1 (#706)
 - **Projectroles**
     - Enforce 3 character minimum limit for terms in multi-term search (#715)
+    - Refactor ``get_backend_include()`` template tag
 
 Fixed
 -----